requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
pydub==0.25.1
google-generativeai==0.3.2
flask-cors==4.0.0
openai-whisper

